Refactor helm init command for reuse, allowing other programs to initialize local helm home directory without shelling out to the helm binary
