+++ cmd/helm/installer/init_windows.go
@@ -16,7 +16,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package main
+package installer // import "k8s.io/helm/cmd/helm/installer"
 
 import (
 	"os"
